Reducir el tamaño del codigo
pasar el main de tener aproximadamente 1000 lineas de codigo a 500, aún es posible reducir más
