+++ Proyecto con interfaz inicial/Prueba/nbproject/private/private.xml
@@ -2,9 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/Fernando/Documents/ReposPdP/Proyecto/Proyecto%20con%20interfaz%20inicial/Prueba/src/prueba/Prueba.java</file>
-            <file>file:/C:/Users/Fernando/Documents/ReposPdP/Proyecto/Proyecto%20con%20interfaz%20inicial/Prueba/src/modelo/Atomos.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
