+++ Proyecto con interfaz inicial/Prueba/src/modelo/Botonesevent.java
@@ -1,0 +1,34 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package modelo;
+
+import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+
+public class Botonesevent {
+    
+    public void molecbtn(Button Molec,Label Nombre){
+        Molec.setOnAction(new EventHandler<ActionEvent>() {
+            @Override
+            public void handle(ActionEvent event) {
+
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setHeaderText(null);
+                alert.setTitle("Info");
+                alert.setContentText("Se a creado la molecula " + Nombre.getText());
+                alert.showAndWait();
+                
+            }
+        });
+        
+        
+    }
+    
+    
+}
